Fixed a flush_taskqueue bug and made the task list singly linked
1. worker thread now removes a task from the task list after it has
processed the same. Previously, the task was being removed before
processing, hence also bringing down the 'num tasks in the list'
counter. So, this counter will reach 0 while the last task is
being processed, and if flush_taskqueue is called at this time, it
returns because the 'num tasks in the list' counter is 0. Not the
desired behaviour. Also, if the main process is just waiting on
flush_taskqueue before exiting, it will exit even before the thread
may finish, thereby killing the last task, in each queue.

2. removed the previous task pointer from the task structure. It had
no use. So, now the task lists are singly linked.