Adding handlers for worker thread exit
If a worker thread gets killed, we need to take care of the following:
1. Remove entries corresponding to this sub taskqueue from all the flush
structures at the parent taskqueue structure.
2. Prevent any one from queueing new tasks to this queue.
3. Handle any processes waiting on the lock of this sub taskqueue
4. Reduce the number of usable sub_taskqueues at the parent taskqueue struct,
and if it becomes 0, gracefully destroy the taskqueue.
5. If the number of usable sub_taskqueues is greater than 0, redistribute the
ill fated sub_taskqueues' tasks to other sub_taskqueues.

Added a function for removing entries from the flush structures(point 1 above)
when a sub_taskqueue is killed. This prevents deadlocks for any waiting
processes.